JunctionSeqCountSet-class {JunctionSeq}	R Documentation

Class "JunctionSeqCountSet"

Description

A JunctionSeqCountSet is a container class that contains all information pertaining to a JunctionSeq analysis and dataset. In general, these methods and slots will not be used by the end-users. In general, JunctionSeqCountSet objects will be created by readJunctionSeqCounts or runJunctionSeqAnalyses and are to be manipulated byhigh-level JunctionSeq functions such as estimateEffectSizes or fitJunctionSeqDispersionFunction.

The methods documented here are for use by advanced users only.

Details

Slots:

designColumns

A character vector with the column names in the design data.frame.

dispFitCoefs

The dispersion fit coefficients.

fittedMu

Fitted mu values generated by DESeq2 code.

dispFunctionType

A list of various variables defining the dispersion functon used.

dispFunction

A function that converts a base mean to a fitted dispersion based on all included count bins.

dispFunctionJct

A function that converts a base mean to a fitted dispersion based only on the splice junction bins.

dispFunctionExon

A function that converts a base mean to a fitted dispersion based only on the exon bins.

formulas

A list of formulas used.

annotationFile

The annotation file.

geneCountData

A matrix of the gene-level counts

countVectors

A matrix of the count vectors.

altSizeFactors

(Not currently used)

plottingEstimates

A list of fitted estimates, for plotting.

plottingEstimatesVST

(Not currently used)

geneLevelPlottingEstimates

A list of gene-level fitted estimates, for plotting.

modelFitForHypothesisTest

(Not currently used)

modelFitForEffectSize

(Not currently used)

flatGffData

A data.frame representation of the flattened gff annotation for each countbin.

flatGffGeneData

A data.frame representation of the flattened gff annotation for each gene.

analysisType

The type of analysis. Character string.

DESeqDataSet

The specially-constructed DESeqDataSet, to be passed to the internally-loaded DESeq2 code.

modelCoefficientsSample:

Object of class "list". Placeholder slot for model coefficients (used for diagnostic testing of code).

modelCoefficientsGene:

Object of class "list". Placeholder slot for model coefficients (used for diagnostic testing of code).

assayData:

Object of class "AssayData". Contains various data.

phenoData:

Object of class "AnnotatedDataFrame". Phenotype data.

featureData:

Object of class "AnnotatedDataFrame". Counting bin data.

experimentData:

Object of class "MIAxE". Information on the experiment.

annotation:

Object of class "character". Not used.

protocolData:

Object of class "AnnotatedDataFrame". Information on the code.

.__classVersion__:

Object of class "Versions". The version of the JunctionSeqCountSet.

Constructor

newJunctionSeqCountSet( countData, geneCountData, design, geneIDs, countbinIDs, featureIntervals=NULL, transcripts=NULL): Creates a new JunctionSeqCountSet

countData

A matrix of junction-level count data of non-negative integer values. The rows correspond to counts for each splice-junction counting bin, the columns correspond to samples. Note that biological replicates should each get their own column, while the counts of technical replicates (i.e., several sequencing runs/lanes from the same sample) should be summed up into a single column.

geneCountData

A matrix of gene-level count data of non-negative integer values. The rows correspond to counts for each gene, the columns correspond to samples. Note that biological replicates should each get their own column, while the counts of technical replicates (i.e., several sequencing runs/lanes from the same sample) should be summed up into a single column. Must have the same dimensions as countData.

design

A data frame consisting of all factors to be included in the analysis. All columns should be factors. Each column should represent a different variable, each row should represent a different sample. The number of rows must equal the number of columns in geneCountData and countData.

geneIDs

A character vector of gene indentifiers for each splice junction. The length must equal the number of rows in countData.

countbinIDs

A character vector of splice-junction-locus indentifiers for each splice junction. The length must equal the number of rows in countData.

featureIntervals

Optional. A data.frame with 4 columns: "chr", "start", "end", and "strand". chr and strand should be character vectors or factors, start and end must be integers.

transcripts

Optional. Character vector listing the transcripts that each splice junction belongs to. Some junctions may belong to more than one transcripts. In this case, transcripts should be separated with the "+" character.

This constructor function SHOULD NOT BE USED in normal operation. Instead you should use the readJunctionSeqCounts function, which returns a new JunctionSeqCountSet.

Extends

Class "eSet", directly. Class "VersionedBiobase", by class "eSet", distance 2. Class "Versioned", by class "eSet", distance 3.

Note

End-users generally will not use any of these slots or methods directly. However, they may be useful for model fit diagnostics and similar statistical experimentation.

You can access method-mode information using the "AltMethods" attribute, and a list of all calls using the "callStack" attribute.

Author(s)

Stephen Hartley

See Also

The proper way to create a JunctionSeqCountSet is to use readJunctionSeqCounts or runJunctionSeqAnalyses.

Examples

showClass("JunctionSeqCountSet")

## Class "JunctionSeqCountSet" [package "JunctionSeq"]
## 
## Slots:
##                                                             
## Name:               designColumns               dispFitCoefs
## Class:                  character                    numeric
##                                                             
## Name:                    fittedMu           dispFunctionType
## Class:                     matrix                       list
##                                                             
## Name:                dispFunction            dispFunctionJct
## Class:                   function                   function
##                                                             
## Name:            dispFunctionExon                   formulas
## Class:                   function                       list
##                                                             
## Name:              annotationFile              geneCountData
## Class:                  character                     matrix
##                                                             
## Name:                countVectors             altSizeFactors
## Class:                     matrix                 data.frame
##                                                             
## Name:           plottingEstimates       plottingEstimatesVST
## Class:                       list                       list
##                                                             
## Name:  geneLevelPlottingEstimates  modelFitForHypothesisTest
## Class:                       list                       list
##                                                             
## Name:       modelFitForEffectSize                flatGffData
## Class:                       list                 data.frame
##                                                             
## Name:             flatGffGeneData               analysisType
## Class:                       list                  character
##                                                             
## Name:                DESeqDataSet    modelCoefficientsSample
## Class:               DESeqDataSet                       list
##                                                             
## Name:       modelCoefficientsGene                  assayData
## Class:                       list                  AssayData
##                                                             
## Name:                   phenoData                featureData
## Class:         AnnotatedDataFrame         AnnotatedDataFrame
##                                                             
## Name:              experimentData                 annotation
## Class:                      MIAxE                  character
##                                                             
## Name:                protocolData          .__classVersion__
## Class:         AnnotatedDataFrame                   Versions
## 
## Extends: 
## Class "eSet", directly
## Class "VersionedBiobase", by class "eSet", distance 2
## Class "Versioned", by class "eSet", distance 3

[Package JunctionSeq version 1.5.4 Index]