fitDispersionFunction {JunctionSeq}	R Documentation

Fit Shared Dispersion Function

Description

Fit dispersion function to share dispersion information between features across the genome.

This function is called internally by the runJunctionSeqAnalyses function, and thus for most purposes users should not need to call this function directly. It may be useful to advanced users performing non-standard analyses.

Usage

  fitDispersionFunction(jscs, fitType = c("parametric","local","mean"), 
                        fitDispersionsForExonsAndJunctionsSeparately = TRUE, 
                        advancedMode = TRUE,
                        verbose = TRUE);

Arguments

jscs
A JunctionSeqCountSet. Usually initially created by readJunctionSeqCounts. Size factors must be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions.

fitType
Which dispersion fit method to use. Not that currently only the parametric (default) method is supported. The other methods are currently nonfunctional and will be added in future releases.

fitDispersionsForExonsAndJunctionsSeparately
When running a "hybrid" analysis in which both exons and splice junctions are being tested simultaniously, this parameter determines whether a single fitted dispersion model should be fitted for both exons and splice junctions, or if separate fitted dispersions should be calculated for each. By default the dispersions are run separately.

advancedMode
Switches between the current default method and an older variant.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

Value

A JunctionSeqCountSet, with dispersion results included.

[Package JunctionSeq version 0.3.40 Index]