generateAllExpressionEstimates {JunctionSeq}	R Documentation

TODO

Description

ADVANCED USERS ONLY: This function is intended for advanced users only. For general use, using the generateCompleteResults function is recommended, which will in turn call this function internally.

Usage

  generateAllExpressionEstimates(jscs,
                                 nCores = 1,
                                 fitExpToVar="condition",
                                 verbose=TRUE)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testJunctionsForDJU.

nCores
The number of cores to use. Note that multicore functionality may not be available on all platforms.

fitExpToVar
Character value. The biological condition to fit the model to.

verbose
Logical value. Determines how much debugging into to message to the console.

[Package JunctionSeq version 0.3.5 Index]