generateCompleteResults {JunctionSeq}	R Documentation

generateCompleteResults (DEPRECIATED)

Description

This function is DEPRECIATED. Use writeCompleteResults instead.

This function takes the raw DEXSeq results and merges in feature annotations, as well as calculating and merging in a number of different normalized and fitted values for each level of the condition variable.

Usage

generateCompleteResults(jscs, outfile.prefix = NULL, 
                        verbose = TRUE,nCores=1,
                        gzip.output = TRUE)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testForDiffUsage.

outfile.prefix
A string indicating the filename prefix where output files should be saved. If NULL, no data will be saved to file. (Default=NULL)

verbose
A boolean flag indicating whether or not to print progress information during execution. (Default=TRUE)

nCores
If the package multicore is installed, the number of cores to use. If multicore is not installed, then setting this option anything other than the default of 1 does nothing but produce a warning.

gzip.output
A boolean indicating whether output files should be gzipped.

Value

A data frame containing the results for all listed splice sites of genes for which the GLM converged successfully.

Each row is a splice site. The columns are:

First, All columns from the res input file.

exprVST_conditionVal: the variance-stabilizing-transformed fitted expression estimates for the splice site at each condition value.

expr_conditionVal: the un-transformed fitted expression estimates for the splice site at each condition value.

rExprVST_conditionVal: the variance-stabilizing-transformed fitted relative expression estimates for the splice site at each condition value.

rExpr_conditionVal: as above, but without the Variance-stabilizing transformation.

normCountVST_sampleID: Variance-stabilizing-transformed normalized counts for the splice site at each sample.

normCount_sampleID: Variance-stabilizing-transformed normalized counts for the splice site at each sample.

Finally, (if anno.file is non-null), the columns from the annotation file, matched to the given genes/splice sites.

Assuming the annotation file is the standard tab-delimited annotation file generated by prepare_annotation_with_splices.py, the final columns will be:

featureType: the type of the feature (for these analyses, always equal to either “splice_site” or “novel_splice_site”.

chrom: The chromosome position.

start

end

strand: ‘+’ or ‘-’

gene_id

part_number: the 3-digit sub-feature ID

transcripts: the list of transcripts, delimited by ‘+’

[Package JunctionSeq version 0.3.40 Index]