plotDispEsts {JunctionSeq}	R Documentation

Plot Fitted and Test-wise Dispersion

Description

Plots the countbin-specific estimated dispersion and the fitted dispersion curve.

Usage

  plotDispEsts( jscs, ymin, 
                linecol="#ff000080",
                xlab = "mean of normalized counts", 
                ylab = "dispersion",
                log = "xy", cex = 0.45, ... )

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testJunctionsForDJU.

ymin
The minimum for the y-axis.

linecol
The line color to use for the fit line.

xlab
The label for the x-axis.

ylab
The label for the y-axis.

log
Logical value. Determines which axes to log-scale.

cex
The cex value to be passed to plot.

...
Additional options to pass to plotting functions, particularly graphical parameters.

[Package JunctionSeq version 0.3.5 Index]