plotJunctionSeqResultsForGene {JunctionSeq}	R Documentation

Plot JunctionSeq results

Description

Creates one results plot for one gene.

Usage

plotJunctionSeqResultsForGene(geneID, jscs, flat.gff.data, 
                              colorRed.FDR.threshold=0.05,
                              plot.type = "expr", 
                              displayTranscripts = FALSE,
                              color = NULL, 
                              use.vst = FALSE, use.log = FALSE,truncateBelowOne = TRUE,
                              exon.rescale.factor = 0.3,
                              label.p.vals = TRUE, 
                              plot.lwd = 3,axes.lwd = plot.lwd, anno.lwd = plot.lwd, 
                              par.cex = 1, anno.cex.text = 1,
                              anno.cex.axis=anno.cex.text, anno.cex.main = anno.cex.text * 1.2,
                              fit.countbin.names = TRUE,
                              debug.mode = FALSE,
                              plot.exon.results = FALSE, plot.splice.results = TRUE, plot.novel.splice.results = plot.splice.results, 
                              plot.untestable.results = FALSE,
                              show.strand.arrows = 20,
                              sort.features = TRUE,
                              drawCoordinates = TRUE,
                              arrows.length = 0.25,
                              title.main=NULL, title.ylab=NULL,
                              verbose=TRUE,
                              ...)

Arguments

geneID
Character string. The gene to the plotted.

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. However in this case a number of additional steps will be necessary: Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testJunctionsForDJU. Effect sizes and parameter estimates must be created via estimateEffectSizes.

flat.gff.data
A data frame containing the annotation data. This can be extracted from a flat gff file using the readAnnotationData function.

colorRed.FDR.threshold
The adjusted-p-value threshold used to determine whether a feature should be marked as "significant" and colored pink. By default this will be the same as the FDR.threshold.

plot.type
Character string. Determines which plot to produce. Options are: "expr" for "expression", or mean normalized read counts by experimental condition, "rExpr" for "relative" expression relative to gene-level expression, "normCounts" for normalized read counts for each sample, and "rawCounts" for raw read counts for each sample.

displayTranscripts
Logical. If true, then the full set of annotated transcripts will be displayed below the expression plot (to a maximum of 42 different TX).

color
A vector of R colors, named for each possible value of condition. By default, it will attempt to choose reasonable colors for each condition.

use.vst
Logical. If TRUE, all plots will be scaled via a variance stabilizing transform.

use.log
Logical. If TRUE, all plots will be log-scaled.

truncateBelowOne
Logical. If TRUE, all values between 0 and 1 will be drawn with a linear scale. If use.log is FALSE, this does nothing.

exon.rescale.factor
Numeric. Exons will be proportionately scaled-up so that the exonic regions make up this fraction of the horizontal plotting area. If negative, exons and introns will be plotted to a common scale.

label.p.vals
Logical. If TRUE, then statistically significant p-values will be labelled.

plot.lwd
the line width for the plotting lines.

axes.lwd
the line width for the axes.

anno.lwd
the line width for the various other annotation lines.

par.cex
The base cex value to be passed to par() immediately before all plots are created. See par.

anno.cex.text
The font size multiplier for most annotation text. This will be multiplied by a factor of the par.cex value. More specifically: The cex value to be passed to all function calls that take graphical parameters. See par.

anno.cex.axis
The font size multiplier for the axis text. This will be multiplied by a factor of the par.cex value. More specifically: The cex.axis value to be passed to all function calls that take graphical parameters. See par.

anno.cex.main
The font size multiplier for the main title text. This will be multiplied by a factor of the par.cex value. More specifically: The cex.main value to be passed to all function calls that take graphical parameters. See par.

fit.countbin.names
Logical. If TRUE, then splice-junction-locus labels should be rescaled to fit in whatever horizontal space is available.

debug.mode
Logical. If TRUE, print additional debugging information during execution.

plot.exon.results
Logical. If TRUE, plot results for exons. Technically speaking, JunctionSeq can be used to do DEXSeq-style analyses on exon partitions. However this functionality is for advanced users only.

plot.splice.results
Logical. If TRUE, plot results for splice junctions. For advanced users only.

plot.novel.splice.results
Logical. If TRUE, plot results for novel splice junctions. If false, novel splice junctions will be ignored.

plot.untestable.results
Logical. If TRUE, plots splice junctions that had coverage that was too low to be tested.

show.strand.arrows
The number of strand-direction arrows to display.

sort.features
Logical. If TRUE, sort features by genomic position.

drawCoordinates
Whether to label the genomic coordinates at the bottom of the plot.

arrows.length
The length of the strand-direction arrows, in inches.

title.main
Character string. Overrides the default main plot title.

title.ylab
Character string. Overrides the default y-axis label.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

...
Additional options to pass to plotting functions, particularly graphical parameters.

[Package JunctionSeq version 0.3.5 Index]