readAnnotationData {JunctionSeq}	R Documentation

Read junctionSeq annotation files produced by QoRTs.

Description

This function reads the "flattened" gff annotation file created by QoRTs. This annotation file contains all the gene, transcript, exon, and junction ID's and their loci. In general this function is not used by the end-user, but is called internally by runJunctionSeqAnalyses or readJunctionSeqCounts.

Usage

   readAnnotationData(flat.gff.file)

Arguments

flat.gff.file
Character string. The filename of the "flat" gff annotation file. The file may be gzip-compressed. This "flat" gff file must be produced by the QoRTs jar utility using the makeFlatGtf or mergeNovelSplices functions (depending on whether inclusion of novel splice junctions is desired).

Value

A data.frame object, containing the annotation information from the flat gff file.

Examples

gff.file <- system.file("extdata/cts/withNovel.forJunctionSeq.gff.gz",
                        package="JctSeqData");

#Parse the GFF file:
annoData <- readAnnotationData(gff.file);
head(annoData);

##               featureName featureType chrom  start    end strand
## 1 ENSRNOG00000050954:E001 exonic_part chr14 182677 183555      -
## 2 ENSRNOG00000050954:E002 exonic_part chr14 184706 184830      -
## 3 ENSRNOG00000050954:E003 exonic_part chr14 186228 186453      -
## 4 ENSRNOG00000050954:E004 exonic_part chr14 189145 189973      -
## 5 ENSRNOG00000050954:E005 exonic_part chr14 190012 190189      -
## 6 ENSRNOG00000050954:E006 exonic_part chr14 190205 190293      -
##              gene_id part_number        transcripts
## 1 ENSRNOG00000050954         001 ENSRNOT00000073973
## 2 ENSRNOG00000050954         002 ENSRNOT00000073973
## 3 ENSRNOG00000050954         003 ENSRNOT00000073973
## 4 ENSRNOG00000050954         004 ENSRNOT00000073973
## 5 ENSRNOG00000050954         005 ENSRNOT00000073973
## 6 ENSRNOG00000050954         006 ENSRNOT00000073973

[Package JunctionSeq version 0.6.42 Index]