readJunctionSeqCounts {JunctionSeq}	R Documentation

Read junctionSeq count files

Description

This function reads count data (usually produced by QoRTs) and compiles them into a JunctionSeqCountSet.

For most purposes this function should not be needed, but might be useful to advanced users performing non-standard analyses. This and all other standard analysis functions can be run automatically via the runJunctionSeqAnalyses function.

Usage

readJunctionSeqCounts(countfiles, 
                      countdata,
                      samplenames,
                      design, 
                      flat.gff.file, 
                      use.splice.sites = TRUE, 
                      use.novel.splice.sites = TRUE, 
                      use.exons = FALSE, 
                      verbose = TRUE)

Arguments

countfiles
Character vector. The filenames of the count files generated by QoRTs. The counts must all be generated using equivalent QoRTs parameters. The strandedness must be the same, as well as the inclusion of novel junctions.

countdata
List. An alternative parameterization. Instead of supplying count files using the countfiles parameter, you can pass a list of data frames, one for each sample. Each data frame should contain two columns: the first should be the feature id and the second should be the counts. This list must have the same length as the samplenames parameter.

samplenames
Character vector. A vector of full sample names, in the same order as the countfiles parameter.

design
A data frame containing the condition variable and all desired covariates. All variables should be factors.

flat.gff.file
Character string. The filename of the "flat" gff annotation file. Can be gzip-compressed. This "flat" gff file must be produced by the QoRTs jar utility using the makeFlatGtf or mergeNovelSplices functions (depending on whether inclusion of novel splice junctions is desired).

NOTE: This option is technically optional, but strongly recommended. If it is not included, then attempts to plot the results will crash unless (non-default) options are used to deactivate the plotting of genomic coordinates and transcript information

use.splice.sites
FOR ADVANCED USERS ONLY: Logical value. If TRUE, splice junctions will be included in the analyses.

use.novel.splice.sites
Logical value. If TRUE, splice junctions that are marked as novel in the gff annotation will be included in the analysis. Note that this is does nothing if there are no novel splice junctions in the count tables.

use.exons
FOR ADVANCED USERS ONLY: Logical value. If TRUE, exon segments will be included in the analyses.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

Value

A JunctionSeqCountSet.

[Package JunctionSeq version 0.3.5 Index]