writeCompleteResults {JunctionSeq}	R Documentation

Produce output data files, given annotation files and DEXSeq exonCountSet object and DEXSeq results data.

Description

This function takes the raw DEXSeq results and merges in feature annotations, as well as calculating and merging in a number of different normalized and fitted values for each level of the condition variable.

Usage

writeCompleteResults(jscs, outfile.prefix, 
                        gzip.output = TRUE,
                        FDR.threshold = 0.01,
                        save.allGenes = TRUE, save.sigGenes = TRUE,
                        save.fit = FALSE, save.VST = FALSE,
                        save.bedTracks = TRUE,
                        save.jscs = FALSE,
                        bedtrack.format = c("BED", "GTF", "GFF3"),
                        verbose = TRUE)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. However in this case a number of additional steps will be necessary: Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testForDiffUsage. Effect sizes and parameter estimates must be created via estimateEffectSizes.

outfile.prefix
A string indicating the filename prefix where output files should be saved.

gzip.output
Logical. If TRUE, then all ".txt" text files should be gzip-compressed to save space.

FDR.threshold
The adjusted-p-value threshold used to determine statistical significance.

save.allGenes
Logical. Whether to save files containing data for all genes.

save.sigGenes
Logical. Whether to save a separate set of files containing data for only the significant genes. If this and save.allGenes are both true then two sets of files will be generated.

save.fit
Logical. Whether to save model fit data.

save.VST
Logical. Whether to save VST-transformed data.

save.bedTracks
Logical. Whether to save "bed" junction coverage tracks.

save.jscs
Logical. Whether to the entire JunctionSeqCountSet object. Default is FALSE.

bedtrack.format
Character string. The format to use for the browser tracks.

verbose
A boolean flag indicating whether or not to print progress information during execution. (Default=FALSE)

Details

Saves a wide variety of data from the analyses.

Value

This is a side-effecting function, and does not return a value.

Examples

data(exampleDataSet,package="JctSeqData");
#Write results tables and browser track files:
writeCompleteResults(jscs, outfile.prefix = "./results.");

## > STARTING writeCompleteResults (v1.5.4) (Thu Mar 30 17:21:26 2017)

## > wcr: outfile.prefix:  ./results.

## > wcr: FDR.threshold:  0.01

## > wcr: save.allGenes:  TRUE

## > wcr: save.sigGenes:  TRUE

## > wcr: save.fit:  FALSE

## > wcr: save.VST:  FALSE

## > wcr: bedtrack.format:  BED

## > wcr: Writing sizeFactors.

## > wcr: Writing results for 117 genes.

## > wcr:     Found 1516 counting bins belonging to these genes.

##    Compiling data table. Thu Mar 30 17:21:27 2017

## > wcr: Writing results for 5 genes with 1 or more significant junctions (at adjusted-p-value threshold 0.01)

## > wcr:     Found 99 counting bins belonging to those genes.

## > wcr:     Writing sigGenes.expression.data.txt

## > wcr:     Writing sigGenes.results.txt

## > wcr:     Writing bed tracks.

## > DONE writeCompleteResults (Thu Mar 30 17:21:27 2017)

## Not run: 
##D ########################################
##D #Set up example data:
##D decoder.file <- system.file(
##D                   "extdata/annoFiles/decoder.bySample.txt",
##D                   package="JctSeqData");
##D decoder <- read.table(decoder.file,
##D                   header=TRUE,
##D                   stringsAsFactors=FALSE);
##D gff.file <- system.file(
##D             "extdata/cts/withNovel.forJunctionSeq.gff.gz",
##D             package="JctSeqData");
##D countFiles <- system.file(paste0("extdata/cts/",
##D      decoder$sample.ID,
##D      "/QC.spliceJunctionAndExonCounts.withNovel.forJunctionSeq.txt.gz"),
##D      package="JctSeqData");
##D ######################
##D #Run example analysis:
##D jscs <- runJunctionSeqAnalyses(sample.files = countFiles,
##D            sample.names = decoder$sample.ID,
##D            condition=factor(decoder$group.ID),
##D            flat.gff.file = gff.file,
##D            analysis.type = "junctionsAndExons"
##D );
##D ########################################
##D 
##D #Write results tables and browser track files:
##D writeCompleteResults(jscs, outfile.prefix = "./results.");
##D 
## End(Not run)

[Package JunctionSeq version 1.5.4 Index]