buildAllPlots {JunctionSeq}	R Documentation

Create and save a full battery of JunctionSeq expression plots.

Description

Saves a large battery of plots displaying the analysis results, for the purposes of data visualization. By default it saves a full set of plots for every gene that shows statistical significance and the adjusted-p < 0.01 level. Alternatively, it can be supplied with a specific gene list using the gene.list parameter, and will plot those specific genes.

Note that this function has MANY parameters, allowing the user to tweak the appearance of the plots to suit their particular needs and preferences. Don't be daunted: the default parameters are probably fine for most purposes.

Usage

  buildAllPlots(jscs,
                outfile.prefix = "./",
                gene.list = NULL, FDR.threshold = 0.01, 
                use.plotting.device = "png",
                use.vst=FALSE,use.log = TRUE, truncateBelowOne = TRUE, 
                exon.rescale.factor = 0.3,
                subdirectories.by.type = TRUE,
                ma.plot=FALSE, variance.plot=FALSE,
                with.TX=TRUE,without.TX=TRUE,
                expr.plot=TRUE,normCounts.plot=TRUE,
                rExpr.plot=FALSE,rawCounts.plot=FALSE,
                colorRed.FDR.threshold = FDR.threshold, 
                color=NULL,
                plot.gene.level.expression = NULL,
                plot.exon.results, plot.junction.results, plot.novel.junction.results,
                plot.untestable.results = FALSE,
                plot.lwd=3, axes.lwd = plot.lwd, anno.lwd = plot.lwd, 
                par.cex = 1, anno.cex.text = 2, anno.cex.axis = anno.cex.text, anno.cex.main = anno.cex.text * 1.2,
                drawCoordinates = TRUE,
                yAxisLabels.inExponentialForm = FALSE,
                show.strand.arrows = 10, arrows.length = 0.125,
                graph.margins = c(2,3,3,2),
                base.plot.height = 22.222, base.plot.width = 22.222, 
                base.plot.units = "in", 
                GENE.annotation.relative.height = 0.2,
                TX.annotation.relative.height = 0.025,
                autoscale.height.to.fit.TX.annotation = TRUE,
                autoscale.width.to.fit.bins = TRUE,
                plotting.device.params = list(), 
                number.plots = TRUE,
                condition.legend.text, include.TX.names = TRUE,
                draw.start.end.sites = TRUE,
                openPlottingDeviceFunc, closePlottingDeviceFunc,
                verbose=TRUE, 
                ...)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. However in this case a number of additional steps will be necessary: Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testForDiffUsage. Effect sizes and parameter estimates must be created via estimateEffectSizes.

outfile.prefix
The prefix file path to save the images to.

gene.list
Character vector. List of genes to plot. Either this variable OR FDR.threshold must be set.

FDR.threshold
If this option is used, genes will be selected for plotting based on the presence of statistically significant junctions. The adjusted-p-value threshold used to determine significance. Only genes containing at least 1 significant feature will be plotted.

use.plotting.device
The plotting device to use.

use.vst
Logical. If TRUE, all plots will be scaled via a variance stabilizing transform.

use.log
Logical. If TRUE, all plots will be log-scaled.

truncateBelowOne
Logical. If TRUE, all values below 1 will be linear-scaled. If use.log is FALSE, this does nothing.

exon.rescale.factor
Floating point numeric value. To improve readability the exons drawn in the coordinate annotation are rescaled by default so that they take up 30 percent of the x axis. This makes the plots easier to read, as exons are usually much smaller than introns and thus a group of clustered exons can be hard to distinguish when plotted on a simple scale. If this value is set to NA then the exons and introns will be drawn on the same linear scale.

subdirectories.by.type
Logical value. If TRUE, then subdirectories will be created in the outfile.prefix directory, containing each plot type.

ma.plot
if TRUE, generate and save a MA plot. A MA-plot is a plot of fold change versus base mean normalized counts.

variance.plot
if TRUE, generate and save a plot of the dispersion as a function of the base mean.

with.TX
if TRUE, save expression plots with the full transcripts printed

without.TX
if TRUE, save expression plots with only the compiled exons printed. Note that if this and with.TX.plot are both TRUE, both versions will be saved seperately.

expr.plot
if TRUE, save an expression plot of the expression parameter estimates for each splice site, for each condition.

normCounts.plot
if TRUE, save an expression plot of the normalized mean counts for each splice site, for each sample.

rExpr.plot
if TRUE, save an expression plot of the expression parameter estimates, relative to gene-wide expression, for each splice site, for each condition.

rawCounts.plot
if TRUE, save an expression plot of the raw counts for each splice site, for each sample. Note that these will never be VST-transformed, even when use.vst == TRUE.

colorRed.FDR.threshold
The adjusted-p-value threshold used to determine whether a feature should be marked as "significant" and colored pink. By default this will be the same as the FDR.threshold.

color
A vector of R colors, named for each possible value of condition. By default, it will attempt to choose reasonable colors for each condition.

plot.gene.level.expression
Logical value. If TRUE, gene-level expression (when applicable) will be plotted beside the sub-element-specific expression in a small seperate plotting box. The gene-level expression will NOT be plotted for the "relative expression" plots (since you can't meaningfully plot gene-level expression relative to itself).

plot.exon.results
Logical. If TRUE, plot results for exons. By default everything that was tested will be plotted.

plot.junction.results
Logical. If TRUE, plot results for splice junctions. By default everything that was tested will be plotted.

plot.novel.junction.results
Logical. If TRUE, plot results for novel splice junctions. If false, novel splice junctions will be ignored. By default everything that was tested will be plotted.

plot.untestable.results
Logical. If TRUE, plots splice junctions that had coverage that was too low to be tested.

plot.lwd
the line width for the plotting lines.

axes.lwd
the line width for the axes.

anno.lwd
the line width for the various other annotation lines.

par.cex
The base cex value to be passed to par() immediately before all plots are created. See par.

anno.cex.text
The font size multiplier for most annotation text. This will be multiplied by a factor of the par.cex value. More specifically: The cex value to be passed to all function calls that take graphical parameters. See par.

anno.cex.axis
The font size multiplier for the axis text. This will be multiplied by a factor of the par.cex value. More specifically: The cex.axis value to be passed to all function calls that take graphical parameters. See par.

anno.cex.main
The font size multiplier for the main title text. This will be multiplied by a factor of the par.cex value. More specifically: The cex.main value to be passed to all function calls that take graphical parameters. See par.

drawCoordinates
Whether to label the genomic coordinates at the bottom of the plot.

yAxisLabels.inExponentialForm
Logical. If TRUE, then the y-axis will be labelled in exponential form.

show.strand.arrows
The number of strand-direction arrows to display.

arrows.length
The length of the strand-direction arrows, in inches.

graph.margins
Numeric vector of length 4. These margins values used (as if for par("mar")) for the main graph. The lower part of the plot uses the same left and right margins.

base.plot.height
The base height of the standard-sized plots. Plots that include the full transcript annotation will be expanded by the height of these additional rows. See the withTxPlot.height.multiplier parameter, below.

base.plot.width
The width of the plots.

base.plot.units
The units of measurement for the plot height and width. Default is px, or pixels.

GENE.annotation.relative.height
The height of the "gene track" displayed underneath the main graph, relative to the height of the main graph. By default it is 20 percent.

TX.annotation.relative.height
For all plots that draw the annotated-transcript set (when the with.TX parameter is TRUE), this sets the height of each transcript, as a fraction of the height of the main graph. By default it is 2.5 percent.

autoscale.height.to.fit.TX.annotation
Plots that include the full transcript annotation generally need to have a larger height in order to maintain readability. By default, all plots that include transcripts will be expanded vertically by the height of the additional transcripts. This maintains the same appearance and aspect ratio of the main graph, but also means that the height of the plot will differ between genes. This parameter can be used to override that behavior if a specific figure size is desired. If TRUE, the base.plot.height will be used as the height of the plot, regardless of how many transcripts are included.

autoscale.width.to.fit.bins
Integer value. JunctionSeq will automatically go to great lengths to autofit the data in a readable way. By default, any plots that have more than 35 plotting columns will be widened linearly to fit the excess columns. This parameter can be used to change that value, or turn it off entirely by setting this parameter to NA.

condition.legend.text
List or named vector of character strings. This optional parameter can be used to assign labels to each condition variable values. It should be a list or named vector with length equal to factor(condition). Each element should be named with one of the values from factor(condition), and should contain the label. They will be listed in this order in the figure legend.

include.TX.names
Logical value. If TRUE, then for the plots that include the annotated transcript, the transcript names will be listed. The labels will be drawn at half the size of anno.cex.text.

plotting.device.params
Additional parameters to be passed to the plotting device.

number.plots
Whether to number each gene in the image names, based on either the order they appear in the input gene.list, or in order of ascending p-values.

draw.start.end.sites
Logical value. If TRUE, then transcript start/end sites will be marked on the main gene annotation.

openPlottingDeviceFunc
An R function. This option can be used to use plotting devices other than the ones directly supported by JunctionSeq. This must be a function that must have 3 parameters: filename, heightMult, and widthMult. It should open the desired plotting device. For advanced users only.

closePlottingDeviceFunc
An R function. This must be used in conjunction with openPlottingDeviceFunc. For most devices, you can just use the function "dev.off". For advanced users only.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

...
Additional options to pass to plotting functions, particularly graphical parameters.

[Package JunctionSeq version 0.3.40 Index]