defaultColorList {JunctionSeq}	R Documentation

JunctionSeq Color Parameters

Description

This data set gives default color parameters for JunctionSeq. Any of these parameters can be overridden by using the colorList parameter in link{buildAllPlots}, link{buildAllPlotsForGene}, or link{plotJunctionSeqResults}.

Usage

JUNCTIONSEQ.DEFAULT.COLOR.LIST

Format

A list.

Details

The parameter colorList must be a named list. Any elements with the names listed below will override the default JunctionSeq colors.

SIG.FEATURE.COLOR The color of vertical lines in the plotting panel, for significant features.

NOSIG.FEATURE.COLOR The color of vertical lines in the plotting panel, for tested but non-significant features.

UNTESTABLE.FEATURE.COLOR The color of vertical lines in the plotting panel, for untestable features.

EXCLUDED.FEATURE.COLOR The color of vertical lines in the plotting panel, for features that were not part of the analysis (for example: splice junctions in an exon-based analysis).

SIG.VERTLINE.COLOR, NOSIG.VERTLINE.COLOR, UNTESTABLE.VERTLINE.COLOR: The colors of the vertical dotted lines in the plotting panel.

SIG.FEATURE.BORDER.COLOR, NOSIG.FEATURE.BORDER.COLOR, UNTESTABLE.FEATURE.BORDER.COLOR, EXCLUDED.FEATURE.BORDER.COLOR: The color for borders of features in the gene and transcript annotation.

SIG.FEATURE.FILL.COLOR, NOSIG.FEATURE.FILL.COLOR, UNTESTABLE.FEATURE.FILL.COLOR, EXCLUDED.FEATURE.FILL.COLOR The colors for exon shading in the gene and transcript annotation.

PLOTTING.LINE.COLORS Character vector. The colors for each condition, in order. Will be shortened to the number of conditions.

[Package JunctionSeq version 0.4.19 Index]