defaultColorList {JunctionSeq}	R Documentation

JunctionSeq Color Parameters

Description

This data set specifies some of the default color and style parameters for JunctionSeq. Any of these parameters can be overridden by using the colorList parameter in buildAllPlots, buildAllPlotsForGene, or plotJunctionSeqResultsForGene.

Usage

JUNCTIONSEQ.DEFAULT.COLOR.LIST

Format

A list.

Details

The parameter colorList must be a named list. Any elements with the names listed below will override the default JunctionSeq colors.

SIG.FEATURE.COLOR The color of vertical lines in the plotting panel, for significant features.

NOSIG.FEATURE.COLOR The color of vertical lines in the plotting panel, for tested but non-significant features.

UNTESTABLE.FEATURE.COLOR The color of vertical lines in the plotting panel, for untestable features.

EXCLUDED.FEATURE.COLOR The color of vertical lines in the plotting panel, for features that were not part of the analysis (for example: splice junctions in an exon-based analysis).

SIG.VERTLINE.COLOR, NOSIG.VERTLINE.COLOR, UNTESTABLE.VERTLINE.COLOR: The colors of the vertical dotted lines in the plotting panel.

SIG.FEATURE.BORDER.COLOR, NOSIG.FEATURE.BORDER.COLOR, UNTESTABLE.FEATURE.BORDER.COLOR, EXCLUDED.FEATURE.BORDER.COLOR: The color for borders of features in the gene and transcript annotation.

SIG.FEATURE.FILL.COLOR, NOSIG.FEATURE.FILL.COLOR, UNTESTABLE.FEATURE.FILL.COLOR, EXCLUDED.FEATURE.FILL.COLOR The colors for exon shading in the gene and transcript annotation.

KNOWN.SPLICE.LTY, NOVEL.SPLICE.LTY: These set the "lty" parameter for the known and novel splice junction diagrams. These must be specified as one of the character specifications. In other words, either a descriptive specification like "solid","dotted",etc., or by an even number of nonzero hexadecimal digits. See the section "Line Type Specification" in the graphical parameters help (linkpar). You CANNOT use the single-digit index specification, as this will fail when used in vector form with the others (technically you could use them, but only if you reset ALL lty parameters).

EXON.CONNECTION.LTY, NOVEL.SPLICE.CONNECTION.LTY, KNOWN.SPLICE.CONNECTION.LTY: The "lty" parameters for the lines that connect the features on the gene schematic diagram to the plotting panel. These have the same requirements as the previous set of lty parameters.

PLOTTING.LINE.COLORS Character vector. The colors for each condition, in order. Will be shortened to the number of conditions.

Value

See above.

Examples

data(exampleDataSet,package="JctSeqData");

#Set a few alternative colors:
buildAllPlotsForGene(geneID = "ENSRNOG00000009281", jscs, 
                     outfile.prefix = "./oddColors.",
                     colorList = list(SIG.FEATURE.COLOR = "red",
                                      SIG.FEATURE.FILL.COLOR = "green",
                                      NOSIG.FEATURE.FILL.COLOR = "blue"
                                      ));

## starting buildAllPlotsForGene() for geneID: ENSRNOG00000009281
## Starting nested heights...
## > pJSRfG(): ENSRNOG00000009281, plot.type: expr (with TX)
## > pJSRfG(): ENSRNOG00000009281, plot.type: expr
## > pJSRfG(): ENSRNOG00000009281, plot.type: normCounts (with TX)
## > pJSRfG(): ENSRNOG00000009281, plot.type: normCounts
## > pJSRfG(): ENSRNOG00000009281, plot.type: rExpr (with TX)
## > pJSRfG(): ENSRNOG00000009281, plot.type: rExpr

[Package JunctionSeq version 1.0.0 Index]