fitDispersionFunction {JunctionSeq}	R Documentation

Fit Shared Dispersion Function

Description

Fit dispersion function to share dispersion information between features across the genome.

For most purposes this function should not be needed, but might be useful to advanced users performing non-standard analyses. This and all other standard analysis functions can be run automatically via the runJunctionSeqAnalyses function.

Usage

  fitDispersionFunction(jscs, verbose);

Arguments

jscs
A JunctionSeqCountSet. Usually initially created by readJunctionSeqCounts. Size factors must be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

Value

A JunctionSeqCountSet, with dispersion results included.

[Package JunctionSeq version 0.3.5 Index]