plotDispEsts {JunctionSeq}	R Documentation

Plot Fitted and Test-wise Dispersion

Description

Plots the countbin-specific estimated dispersion and the fitted dispersion curve.

Usage

  plotDispEsts(jscs, ylim, xlim, linecol = c("#ff000080","#ff990080"), 
               xlab = "mean of normalized counts", 
               ylab = "dispersion", 
               miniTicks = TRUE, 
               cex = 0.45, pch = c(1,4),
               use.smoothScatter = FALSE, smooth.nbin = 512, nrpoints = 100, 
               plot.exon.results = TRUE, plot.junction.results = TRUE,
               ...)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testForDiffUsage.

ylim
The plotting range for the y-axis.

xlim
The plotting range for the x-axis.

linecol
Character vector of length 2. The line color to use for the fit line. If the fits were performed separately for exons and junctions, the junction line will be drawn with the second color (orange, by default).

xlab
The label for the x-axis.

ylab
The label for the y-axis.

miniTicks
Whether or not to plot smaller ticks at the tenth-decades.

cex
The cex value to be passed to plot, determining the size of the plotted points.

pch
Numeric vector of length 2. Contains the pch parameters for the exon features and the junction features. This determines the character used in plotting. By default the exons are plotted as circles, and junctions are plotted as X's.

use.smoothScatter
Logical. If TRUE, features will be ploted with density shading rather than having each point plotted.

smooth.nbin
The number of bins to smooth, for the density plot, if use.smoothScatter is TRUE.

nrpoints
The number of extra points to plot, if use.smoothScatter is TRUE.

plot.exon.results
Logical. If TRUE, plot results for exons. Technically speaking, JunctionSeq can be used to do DEXSeq-style analyses on exon partitions. However this functionality is for advanced users only.

plot.junction.results
Logical. If TRUE, plot results for splice junctions. For advanced users only.

...
Additional options to pass to plotting functions, particularly graphical parameters.

[Package JunctionSeq version 0.3.41 Index]