readJunctionSeqCounts {JunctionSeq}	R Documentation

Read junctionSeq count files

Description

This function loads read-count data (usually produced by QoRTs) and compiles them into a JunctionSeqCountSet object.

This function is called internally by the runJunctionSeqAnalyses function, and thus for most purposes users should not need to call this function directly. It may be useful to advanced users performing non-standard analyses.

Usage

readJunctionSeqCounts(countfiles, countdata,
                      samplenames, design, 
                      flat.gff.file, 
                      analysis.type = c("hybrid","junctionsOnly","exonsOnly"),
                      nCores = 1,
                      use.exons, use.junctions, 
                      use.novel.junctions = TRUE, 
                      verbose = TRUE)

Arguments

countfiles
Character vector. The filenames of the count files generated by QoRTs. The counts must all be generated using equivalent QoRTs parameters. The strandedness must be the same, as well as the inclusion of novel junctions.

countdata
List. An alternative parameterization. Instead of supplying count files using the countfiles parameter, you can pass a list of data frames, one for each sample. Each data frame should contain two columns: the first should be the feature id and the second should be the counts. This list must have the same length as the samplenames parameter.

samplenames
Character vector. A vector of full sample names, in the same order as the countfiles parameter.

design
A data frame containing the condition variable and all desired covariates. All variables should be factors.

flat.gff.file
Character string. The filename of the "flat" gff annotation file. Can be gzip-compressed. This "flat" gff file must be produced by the QoRTs jar utility using the makeFlatGtf or mergeNovelSplices functions (depending on whether inclusion of novel splice junctions is desired).

NOTE: This option is technically optional, but strongly recommended. If it is not included, then attempts to plot the results will crash unless (non-default) options are used to deactivate the plotting of genomic coordinates and transcript information

analysis.type
Character string. One of "junctionsAndExons", "junctionsOnly", or "exonsOnly". This parameter determines what type of analysis is to be performed. By default JunctionSeq tests both splice junction loci and exonic regions for differential usage (a "hybrid" analysis). This parameter can be used to limit analyses specifically to either splice junction loci or exonic regions.

nCores
The number of cores to use. Note that multicore functionality may not be available on all platforms.

use.exons
Logical value. This is an alternate parameterization of the analysis.type parameter. If TRUE, then exonic region loci will be included in the analyses and will be tested for differential usage. If this parameter is set, then parameter use.junctions must also be set.

use.junctions
Logical value. This is an alternate parameterization of the analysis.type parameter. If TRUE, then splice junction loci will be included in the analyses and will be tested for differential usage. If this parameter is set, then parameter use.exons must also be set.

use.novel.junctions
Logical value. If TRUE, then novel splice junctions will not be filtered out prior to analysis.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

Value

A JunctionSeqCountSet.

[Package JunctionSeq version 0.3.40 Index]