testForDiffUsage {JunctionSeq}	R Documentation

Test Junctions for Differential Junction Usage

Description

This function runs the hypothesis tests for differential junction usage.

This function is called internally by the runJunctionSeqAnalyses function, and thus for most purposes users should not need to call this function directly. It may be useful to advanced users performing non-standard analyses.

Usage

testForDiffUsage( jscs, 
          test.formula0 = formula(~ sample + countbin), 
          test.formula1 = formula(~ sample + countbin + condition : countbin),
          dispColumn="dispersion", nCores=1, 
          use.alternate.method = TRUE, 
          verbose = TRUE)

Arguments

jscs
A JunctionSeqCountSet. Usually initially created by readJunctionSeqCounts. Dispersions and size factors must be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions.

test.formula0
The formula for the null hypothesis. Note that the condition to be tested must be named "condition".

test.formula1
The formula for the alternative hypothesis. Note that the condition to be tested must be named "condition".

dispColumn
Character value. The name of the fData(jscs) column in which the model dispersion is stored.

nCores
The number of cores to use. Note that multicore functionality may not be available on all platforms. JunctionSeq attempts to use the BiocParallell package if it can be found installed. Otherwise it will attempt to fallback to the multicore package. If neither package can be found it will fallback to single core execution.

use.alternate.method
ADVANCED USERS ONLY: Logical value. Determines whether to use the JunctionSeq standard model framework, or the DEXSeq model framework.

verbose
if TRUE, send debugging and progress messages to the console / stdout.

Value

A JunctionSeqCountSet, with hypothesis test results included.

[Package JunctionSeq version 0.3.40 Index]