writeCompleteResults {JunctionSeq}	R Documentation

Produce output data files, given annotation files and DEXSeq exonCountSet object and DEXSeq results data.

Description

This function takes the raw DEXSeq results and merges in feature annotations, as well as calculating and merging in a number of different normalized and fitted values for each level of the condition variable.

Usage

writeCompleteResults(jscs, outfile.prefix, 
                        gzip.output = TRUE,
                        FDR.threshold = 0.05,
                        save.allGenes = TRUE, save.sigGenes = TRUE,
                        save.fit = FALSE, save.VST = FALSE,
                        save.bedTracks = TRUE,
                        save.jscs = FALSE,
                        verbose = TRUE)

Arguments

jscs
A JunctionSeqCountSet. Usually created by runJunctionSeqAnalyses.

Alternatively, this can be created manually by readJunctionSeqCounts. However in this case a number of additional steps will be necessary: Dispersions and size factors must then be set, usually using functions estimateSizeFactors and estimateJunctionSeqDispersions. Hypothesis tests must be performed by testForDiffUsage. Effect sizes and parameter estimates must be created via estimateEffectSizes.

outfile.prefix
A string indicating the filename prefix where output files should be saved.

gzip.output
Logical. If TRUE, then all ".txt" text files should be gzip-compressed to save space.

FDR.threshold
The adjusted-p-value threshold used to determine statistical significance.

save.allGenes
Logical. Whether to save files containing data for all genes.

save.sigGenes
Logical. Whether to save a separate set of files containing data for only the significant genes. If this and save.allGenes are both true then two sets of files will be generated.

save.fit
Logical. Whether to save model fit data.

save.VST
Logical. Whether to save VST-transformed data.

save.bedTracks
Logical. Whether to save "bed" junction coverage tracks.

save.jscs
Logical. Whether to the entire JunctionSeqCountSet object. Default is FALSE.

verbose
A boolean flag indicating whether or not to print progress information during execution. (Default=FALSE)

Details

Saves a wide variety of data from the analyses.

[Package JunctionSeq version 0.3.41 Index]